+++ sdk/highlight-py/pyproject.toml
@@ -11,6 +11,11 @@ python = "^3.10"
 opentelemetry-api = "^1.15.0"
 opentelemetry-sdk = "^1.15.0"
 opentelemetry-proto = "^1.15.0"
+flask = "^2.2.2"
+opentelemetry-exporter-otlp-proto-http = "^1.15.0"
+opentelemetry-instrumentation = "^0.36b0"
+opentelemetry-distro = {extras = ["otlp"], version = "^0.36b0"}
+opentelemetry-instrumentation-logging = "^0.36b0"
 
 
 [build-system]
